# Conflicts:
#	.idea/.idea.Fruits-Clicker/.idea/workspace.xml
#	Assets/Prefabs/AirText.prefab
#	Assets/Scenes/Game.unity
#	Assets/Scripts/BuyUpgrades/BuyAwakingController/BuyAwaking.cs
#	Assets/Scripts/BuyUpgrades/FruitsController/BuyFruit.cs
#	Assets/Scripts/BuyUpgrades/RebirthController/BuyRebirth.cs
#	Assets/Scripts/BuyUpgrades/RebirthController/RebirthBoostMulti.cs
#	Assets/Scripts/BuyUpgrades/ValutesUpgrades/BuyUpgrades.cs
#	Assets/Scripts/BuyUpgrades/ValutesUpgrades/UpgradesForCoinsController.cs
#	Assets/Scripts/BuyUpgrades/ValutesUpgrades/UpgradesForDimondsController.cs
#	Assets/Scripts/BuyUpgrades/ValutesUpgrades/UpgradesForMultiFruitCoinsController.cs
#	Assets/Scripts/ConvetersOfTexts/ListOfFruitsController/ConvertTextsOfInformPanel.cs
#	Assets/Scripts/ConvetersOfTexts/TextsForRebirthPanelController/ConvertTextForRebirth.cs
#	Assets/Scripts/ConvetersOfTexts/UpgradeTextsController/TextMaxControllerForUpgrades.cs
#	Assets/Scripts/ConvetersOfTexts/UpgradeTextsController/TextsAndSpritesOfUpgradesControll.cs
#	Assets/Scripts/ConvetersOfTexts/ValueTextControllers/ConvertAnyValueToText.cs
#	Assets/Scripts/ConvetersOfTexts/ValueTextControllers/ConvertMultiFruitCoinsAfterPresigeToText.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/ConvertingMultiFruitCoins.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/CreateLargeValuesController.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/FruitCoinsValuteController.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/FruitDimondsValueController.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/MultiFruitCoinsValueController.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/UpgradesController.cs
#	Assets/Scripts/GetResourses/ScriptsForAllValues/ValuteController.cs
#	Assets/Scripts/GetResourses/ScriptsForSystemOfLevelFruits/GetLevelForSuchFruit.cs
#	Assets/Scripts/PanelsControllers/ScriptsOfFlyingText/ControllOfMotionOfTexts.cs
#	Assets/Scripts/PanelsControllers/ScriptsOfFlyingText/CreateATextsOfFruitCoins.cs
#	Assets/Sprites/BuyButton.png.meta
#	Assets/Sprites/Алмазы +.png.meta
#	Assets/Sprites/Банан.png.meta
#	Assets/Sprites/Без имени-1 1.png.meta
#	Assets/Sprites/Без имени-1 2.png.meta
#	Assets/Sprites/Без имени-1 3.png.meta
#	Assets/Sprites/Без имени-1 4.png.meta
#	Assets/Sprites/Без имени-1.png.meta
#	Assets/Sprites/Без имени-2 1.png.meta
#	Assets/Sprites/Без имени-2.png.meta
#	Assets/Sprites/Без имени-3.png.meta
#	Assets/Sprites/Больше алмазных фруктов.png.meta
#	Assets/Sprites/Витамины.png.meta
#	Assets/Sprites/Дополнительные кнопки.png.meta
#	Assets/Sprites/Замок.png.meta
#	Assets/Sprites/Золотые и алмазные фрукты.png.meta
#	Assets/Sprites/Калий.png.meta
#	Assets/Sprites/Капля.png.meta
#	Assets/Sprites/Кнопка 1.png.meta
#	Assets/Sprites/Кнопка 2.png.meta
#	Assets/Sprites/Кнопка закрыть.png.meta
#	Assets/Sprites/Кнопка покупки.png.meta
#	Assets/Sprites/Кнопка.png.meta
#	Assets/Sprites/Кран.png.meta
#	Assets/Sprites/Логотип Игры.png.meta
#	Assets/Sprites/Монетка.png.meta
#	Assets/Sprites/Монетки +.png.meta
#	Assets/Sprites/МультиМонетка.png.meta
#	Assets/Sprites/Панель для таймера 1.png.meta
#	Assets/Sprites/Панель для таймера 2.png.meta
#	Assets/Sprites/Панель для таймера 3.png.meta
#	Assets/Sprites/Панель для таймера.png.meta
#	Assets/Sprites/Повышение уровня яблока.png.meta
#	Assets/Sprites/Спелость.png.meta
#	Assets/Sprites/Стрелка Апгрейда.png.meta
#	Assets/Sprites/Таймер.png.meta
#	Assets/Sprites/Фон для фруктов.png.meta
#	Assets/Sprites/Фруктовый Алмаз.png.meta
#	Assets/Sprites/Фрукты.png.meta
#	Assets/Sprites/Яблоки(золотые или алмазные).png.meta
#	Assets/Sprites/Яблоко.png.meta

+++ Assets/Scripts/PanelsControllers/ScriptsOfFlyingText/CreateATextsOfFruitCoins.cs
@@ -17,7 +17,7 @@ public class CreateATextsOfFruitCoins : MonoBehaviour
    {
       ControllOfMotionOfTexts TextClone = Instantiate(Text.gameObject).GetComponent<ControllOfMotionOfTexts>();
       TextClone.transform.SetParent(Spawner.transform);
-      TextClone.StartMotion((int) FruitCoins.MainMulti);
+      TextClone.StartMotion(FruitCoins.MainMulti);
       Destroy(TextClone.gameObject, 3.5f);
    }
